Moved vmass tool
