+++ Arma3ObjectBuilder/__init__.py
@@ -38,6 +38,7 @@ from . import props
 from . import ui
 
 from . import tool_outliner
+from . import tool_mass
 from . import tool_materials
 from . import tool_hitpoint
 from . import tool_renaming
@@ -351,7 +352,7 @@ modules = (
     ui.import_export_asc,
     ui.import_export_paa,
     tool_outliner,
-    ui.tool_mass,
+    tool_mass,
     tool_materials,
     tool_hitpoint,
     tool_renaming,
