+++ Arma3ObjectBuilder/utilities/__init__.py
@@ -1,4 +1,3 @@
 # In order of dependency
 from . import proxy
-from . import masses
 from . import structure
